Creation of the "pages" package
Type: refactor

Description: Created the "pages" package that will contains the pages's stateful widgets of our app,
then moved the LoginPage widget into it's own login_page.dart file and putted it inside this new package.
