+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:doggies_app/pages/login_page.dart';
 
 void main() {
   runApp(DoggiesApp());
@@ -13,25 +14,6 @@ class DoggiesApp extends StatelessWidget {
         primarySwatch: Colors.blue,
       ),
       home: LoginPage(title: 'Login'),
-    );
-  }
-}
-
-class LoginPage extends StatefulWidget {
-  LoginPage({Key key, this.title}) : super(key: key);
-
-  final String title;
-
-  @override
-  _LoginPageState createState() => _LoginPageState();
-}
-
-class _LoginPageState extends State<LoginPage> {
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: <Widget>[],
     );
   }
 }
